api_service: Add write-only image API stub code.
